@@ -1,35 +1,44 @@
 {
-	"name": "isobit-ui",
-	"version": "0.3.7",
-	"type": "module",
-	"description": "Vue component to play videos",
-	"keywords": [
-		"ui",
-		"table",
-		"autocomplete",
-		"vue"
-	],
-	"author": "Erik Alarcon Pinedo",
-	"license": "MIT",
-	"private": false,
-	"main": "./dist/index.js",
-	"scripts": {
-		"dev": "vue serve v-video-player.vue",
-		"bili": "bili",
-		"eslint": "eslint",
-		"build": "bili --name index --plugin vue --vue.css false",
-		"patch": "git add . & git commit -m \"x\" & git push & bili & npm version patch & npm publish"
-	},
-	"files": [
-		"dist/*"
-	],
-	"devDependencies": {
-		"bili": "^5.0.5",
-		"eslint": "^8.36.0",
-		"rollup-plugin-vue": "^5.1.9",
-		"vue-template-compiler": "^2.6.12"
-	},
-	"dependencies": {
-		"vue": "^2.6.12"
-	}
+  "name": "isobit-ui",
+  "version": "1.0.30",
+  "description": "",
+  "type": "module",
+  "main": "dist/library.js",
+  "module": "dist/library.mjs",
+  "types": "dist/types/index.d.ts",
+  "files": [
+    "dist/*"
+  ],
+  "scripts": {
+    "build": "rollup -c",
+    "dev": "vue serve v-video-player.vue",
+    "eslint": "eslint",
+    "build2": "bili --name index --plugin vue --vue.css false",
+    "patch": "npm version patch & git add . & git commit -m \"x\" & git push & build &  npm publish"
+  },
+  "author": "",
+  "license": "ISC",
+  "devDependencies": {
+    "@rollup/plugin-terser": "^0.4.4",
+    "@rollup/plugin-typescript": "^11.1.6",
+    "rollup": "^4.13.2",
+    "rollup-plugin-css-only": "^4.5.2",
+    "rollup-plugin-peer-deps-external": "^2.2.4",
+    "rollup-plugin-vue": "^6.0.0",
+    "typescript": "^5.4.3",
+    "vue": "^3.2.37",
+    "vue-dts-gen": "^0.3.0",
+    "vue-router": "^4.0.16"
+  },
+  "dependencies": {
+    "@fortawesome/free-regular-svg-icons": "^6.7.2",
+    "@fortawesome/free-solid-svg-icons": "^6.7.2",
+    "@fortawesome/vue-fontawesome": "^3.0.8",
+    "axios": "^1.6.8",
+    "tslib": "^2.6.2"
+  },
+  "peerDependencies": {
+    "ol": "^10.4.0",
+    "pinia": "^2.3.1"
+  }
 }
