@@ -9,6 +9,7 @@
 </template>
 <script>
 export default {
+	name: 'VDataview',
     props: {
         value: Array,
         filterKey: String,
@@ -156,7 +157,7 @@ export default {
 	}
 }
 </script>
-<style>
+<!-- <style>
 	.v-dataview > .v-no-results{
 		padding: 15px;
 		color: rgb(255 193 7);
@@ -178,4 +179,4 @@ export default {
 		border-style:solid;
 		border-width: 1px;
 	}
-</style>
+</style> -->
