@@ -1,61 +1,65 @@
 <template>
-    <div v-show="ready&&visible0" class="ol-unselectable ol-control">
-		<slot></slot>
-		<button v-if="!slotUsed" class="v-map-control-button" v-on:click.prevent="enter">
-			<i v-if="icon" class="fa fa-xs" v-bind:class="icon"></i>
-		</button>
-	</div>
+    <div ref="el" v-show="ready && true" class="ol-unselectable ol-control">
+        <slot></slot>
+        <button v-if="!slotUsed" class="v-map-control-button" @click.prevent="enter">
+            <i v-if="icon" class="fa fa-xs" :class="icon"></i>
+        </button>
+    </div>
 </template>
 <script>
+import Control from 'ol/control/Control';
+import { inject, ref } from "vue";
 export default {
-	props: {
+    name: 'VMapControl',
+    setup(props, { emit }) {
+        //console.log(props);
+        const { visible } = props;
+        const el = ref(null);
+        const visible0 = ref(!!visible);
+        const collector = inject('collector');
+        const ready = ref(null);
+        //me.slotUsed=!!this.$slots.default;
+        const addControl = (map) => {
+            map.addControl(new Control({ element: el.value }));
+            ready.value = true;
+        }
+        collector.push(addControl);
+        const timeout = ref();
+        const enter = (e) => {
+            e.stopPropagation()
+            if (!timeout.value) {
+                emit('click', e);
+            }else{
+                clearTimeout(timeout.value)
+            }
+            timeout.value = setTimeout(() => {
+                timeout.value = null;
+            }, 500)
+        }
+        return { el, ready, visible0, enter }
+    },
+    props: {
         icon: String,
-		visible:{
-			default: 1
-		},
+        visible: {
+            default: 1
+        },
     },
     computed: {
         slotPassed() {
             return !!this.$slots.default;
         }
     },
-    data(){return {ready:null,visible0:null,slotUsed:false}},
-	updated(){
-		var l=this.$el.querySelectorAll(":scope > :not(.v-map-control-button)").length;
-		if(l==0)l=this.$el.textContent.trim().length;
-		this.slotUsed=l;
-		//console.log(this.$slots);
-	},
-	created(){this.visible0=!!this.visible;},
-    mounted(){
-        var me = this;
-		me.slotUsed=!!this.$slots.default;
-		//onsole.log('slotUsed='+me.slotUsed);
-        if(me.$parent.map){
-            me.$parent.map.addControl(new ol.control.Control({element:me.$el}));
-            me.ready=true;
-        }else
-            me.$parent.$on('build', (e)=> {
-                e.map.addControl(new ol.control.Control({element:me.$el}));
-                me.ready=true;
-            });
-			/*
-			var me = this;
-		this.svg=!(this.icon&&this.icon.startsWith('fa-'));
-        me.$parent.$on('build', function (m) {
-            m.addControl(new ol.control.Control({element:me.$el}));
-            me.ready=true;
-        });
-			*/
+    data() { return { slotUsed: false } },
+    updated() {
+        var l = this.$el.querySelectorAll(":scope > :not(.v-map-control-button)").length;
+        if (l == 0) l = this.$el.textContent.trim().length;
+        this.slotUsed = l;
+        //console.log(this.$slots);
     },
-    methods:{
-        enter(e){
-            if(e.x)this.$emit('click',e);
-        },
-		toggle(){
-			this.visible0=!!!this.visible0;
-		}
+    methods: {
+        toggle() {
+            this.visible0 = !!!this.visible0;
+        }
     }
 };
 </script>
-
